feat: album tests 14/14

+++ src/album/album.controller.ts
@@ -1,3 +1,4 @@
+// src/album/album.controller.ts
 import {
   Controller,
   Get,
@@ -6,9 +7,16 @@ import {
   Delete,
   Param,
   Body,
+  HttpCode,
+  HttpStatus,
+  BadRequestException,
+  NotFoundException,
+  UsePipes,
+  ValidationPipe,
 } from '@nestjs/common';
 import { AlbumService } from './album.service';
 import { CreateAlbumDto } from './dto/create-album.dto';
+import { isUUID } from 'class-validator';
 
 @Controller('album')
 export class AlbumController {
@@ -21,21 +29,40 @@ export class AlbumController {
 
   @Get(':id')
   getAlbumById(@Param('id') id: string) {
-    return this.albumService.getAlbumById(id);
+    if (!isUUID(id)) {
+      throw new BadRequestException('Invalid ID format');
+    }
+    const album = this.albumService.getAlbumById(id);
+    if (!album) throw new NotFoundException();
+    return album;
   }
 
   @Post()
+  @UsePipes(new ValidationPipe({ whitelist: true, forbidNonWhitelisted: true }))
+  @HttpCode(HttpStatus.CREATED)
   createAlbum(@Body() createAlbumDto: CreateAlbumDto) {
     return this.albumService.createAlbum(createAlbumDto);
   }
 
   @Put(':id')
+  @UsePipes(new ValidationPipe({ whitelist: true, forbidNonWhitelisted: true }))
   updateAlbum(@Param('id') id: string, @Body() updateAlbumDto: CreateAlbumDto) {
-    return this.albumService.updateAlbum(id, updateAlbumDto);
+    if (!isUUID(id)) {
+      throw new BadRequestException('Invalid ID format');
+    }
+    const album = this.albumService.updateAlbum(id, updateAlbumDto);
+    if (!album) throw new NotFoundException();
+    return album;
   }
 
   @Delete(':id')
+  @HttpCode(HttpStatus.NO_CONTENT)
   deleteAlbum(@Param('id') id: string) {
-    return this.albumService.deleteAlbum(id);
+    if (!isUUID(id)) {
+      throw new BadRequestException('Invalid ID format');
+    }
+    const album = this.albumService.getAlbumById(id);
+    if (!album) throw new NotFoundException();
+    this.albumService.deleteAlbum(id);
   }
 }
